requirements.txt
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-auth>=2.29.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
joblib>=1.4.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.2.0
python-dateutil>=2.9.0
pytz>=2024.1
jinja2>=3.1.0
ujson>=5.10.0